requirements.txt
pandas>=1.1.5
numpy>=1.19.5
scikit-learn>=0.24.2
matplotlib>=3.3.4
joblib>=1.0.1
optuna>=2.10.0
pyyaml>=5.4.1
lightgbm>=3.3.2
catboost>=1.0.6